requirements.txt
starlette
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
requests==2.32.5
python-dotenv==1.2.1
fastapi==0.121.2
uvicorn==0.38.0
